Fix conflict with past actions that failed.
If an action failed, the enabled field would still be 1.
Then after saving a new action for the post in the future,
the new action args would be saved as enabled = 1. Having
Two records enabled for the post, only the first one would
be retrieved, on this case, the older one. Making it always
display the data from the past action that failed instead
of the new data.